[RPL] Fix untrusted loop bound (Coverity issue)
Match the type of the index and the size

+++ Silicon/CommonSocPkg/Library/BootGuardLibCBnT/BootGuardTpmEventLogLib.c
@@ -798,7 +798,7 @@ CaculateDetailPCRExtendValue (
   UINT16                           KeyModulusSize;
   UINT16                           IbbDigestCount;
   UINT16                           IbbDigestSize;
-  UINT8                            Idx;
+  UINT16                           Idx;
   HASH_LIST                        *IbbHashPtr;
 
   AcmSigSize = 0;
